refactor(tools): separate file operation handlers
